Move app keybindings to a dedicated component
